Correccion de errores, configuracion de JSF y formularios

+++ negocio/src/main/java/co/edu/uniquindio/unicine/dto/InfoCompraDto.java
@@ -17,6 +17,6 @@ public class InfoCompraDto {
     private float precioTotal;
     private LocalDateTime fechaCompra;
     private Funcion funcion;
-    private Integer precioBoletas;
-    private Integer precioConfiteria;
+    private double precioBoletas;
+    private double precioConfiteria;
 }
